+++ helper/EsiHelper.php
@@ -209,16 +209,18 @@ class EsiHelper {
 		$transientName = \sanitize_title('eve-esi-data_' . $route);
 		$data = CacheHelper::getTransientCache($transientName);
 
-		if($data === false) {
+		if($data === false || empty($data)) {
 			$data = RemoteHelper::getInstance()->getRemoteData($this->esiUrl . $route);
 
 			/**
 			 * setting the transient caches
 			 */
-			CacheHelper::setTransientCache($transientName, $data, $cacheTime);
+			if(!isset($data->error) && !empty($data)) {
+				CacheHelper::setTransientCache($transientName, $data, $cacheTime);
+			}
 		} // END if($data === false)
 
-		if(!empty($data)) {
+		if(!empty($data) && !isset($data->error)) {
 			$returnValue = \json_decode($data);
 		} // END if(!empty($data))